+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Version 1.0.0 - 2017-??-??
+# Version 1.0.0 - 2017-03-11
  * recording: Support GNOME Shell screencast DBUS service. Allows recording
    under GNOME Shell with XWayland (#33)
  * recording: Support WebM and MP4 as output format (#73)